+++ packages/cinterop/src/jvm/kotlin/io/realm/kotlin/internal/interop/sync/CoreSubscriptionSetState.kt
@@ -22,7 +22,7 @@ import io.realm.kotlin.internal.interop.realm_flx_sync_subscription_set_state_e
 actual enum class CoreSubscriptionSetState(override val nativeValue: Int) : NativeEnumerated {
     RLM_SYNC_SUBSCRIPTION_UNCOMMITTED(realm_flx_sync_subscription_set_state_e.RLM_SYNC_SUBSCRIPTION_UNCOMMITTED),
     RLM_SYNC_SUBSCRIPTION_PENDING(realm_flx_sync_subscription_set_state_e.RLM_SYNC_SUBSCRIPTION_PENDING),
-    RLM_SYNC_BOOTSTRAPPING(realm_flx_sync_subscription_set_state_e.RLM_SYNC_BOOTSTRAPPING),
+    RLM_SYNC_SUBSCRIPTION_BOOTSTRAPPING(realm_flx_sync_subscription_set_state_e.RLM_SYNC_SUBSCRIPTION_BOOTSTRAPPING),
     RLM_SYNC_SUBSCRIPTION_COMPLETE(realm_flx_sync_subscription_set_state_e.RLM_SYNC_SUBSCRIPTION_COMPLETE),
     RLM_SYNC_SUBSCRIPTION_ERROR(realm_flx_sync_subscription_set_state_e.RLM_SYNC_SUBSCRIPTION_ERROR),
     RLM_SYNC_SUBSCRIPTION_SUPERSEDED(realm_flx_sync_subscription_set_state_e.RLM_SYNC_SUBSCRIPTION_SUPERSEDED);